ISSUE-24: add test for shuffle
