+++ genomeshuffle/unittest/CMakeLists.txt
@@ -6,5 +6,8 @@ include_directories(lib/googletest-release-1.8.0/googlemock/include)
 include_directories(/Users/mitsuki/sandbox/genome/genomeshuffle/src)
 include_directories (/Users/mitsuki/sandbox/genome/helper)
 
-add_executable(runTest GeneticCodeTest.cpp test.cpp)
+add_executable(runTest test.cpp)
 target_link_libraries(runTest gtest gtest_main)
+
+add_executable(runGeneticCodeTest GeneticCodeTest.cpp)
+target_link_libraries(runGeneticCodeTest gtest gtest_main)
